refactor(VisualizationManager): simplify layout management and enhance visualization handling
